+++ api/tests/identity-access-management/unit/domain/usecases/authenticate-application_test.js
@@ -1,11 +1,11 @@
 import sinon from 'sinon';
 
-import { PasswordNotMatching } from '../../../../../src/identity-access-management/domain/errors.js';
 import { ApplicationAccessToken } from '../../../../../src/identity-access-management/domain/models/ApplicationAccessToken.js';
 import { authenticateApplication } from '../../../../../src/identity-access-management/domain/usecases/authenticate-application.js';
 import {
   ApplicationScopeNotAllowedError,
   ApplicationWithInvalidCredentialsError,
+  PasswordNotMatching,
 } from '../../../../../src/shared/domain/errors.js';
 import { expect } from '../../../../test-helper.js';
 import { domainBuilder } from '../../../../tooling/domain-builder/domain-builder.js';
