+++ api/tests/shared/unit/domain/services/crypto-service_test.js
@@ -1,7 +1,7 @@
 import bcrypt from 'bcrypt';
 import omit from 'lodash/omit.js';
 
-import { PasswordNotMatching } from '../../../../../src/identity-access-management/domain/errors.js';
+import { PasswordNotMatching } from '../../../../../src/shared/domain/errors.js';
 import { cryptoService } from '../../../../../src/shared/domain/services/crypto-service.js';
 import { expect } from '../../../../test-helper.js';
 import { catchErr } from '../../../../tooling/test-utils/error.js';
